pyproject.toml
[project]
name = "agent-penny"
version = "0.1.0"
description = "Agent Penny is a personal AI assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.27.0",
    "chainlit>=2.9.6",
    "en-core-web-sm",
    "fake-useragent>=2.2.0",
    "faster-whisper>=1.2.1",
    "google-api-python-client>=2.188.0",
    "logfire[aiohttp-client,httpx,requests]>=4.24.0",
    "loguru>=0.7.3",
    "markitdown[pdf]>=0.1.4",
    "perplexityai>=0.26.0",
    "piper-tts>=1.4.2",
    "pydantic-ai-skills>=0.6.0",
    "pydantic-ai-slim[anthropic,bedrock,duckduckgo,google,openai]>=1.71.0",
    "python-slugify>=8.0.4",
    "silero-vad>=6.2.0",
    "tavily-python>=0.7.21",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
    "ua-parser>=1.0.1",
]

[dependency-groups]
dev = [
    "google-api-python-client-stubs>=1.31.0",
    "google-auth-oauthlib>=1.2.4",
    "mypy[faster-cache]>=1.19.1",
    "pre-commit>=4.5.1",
    "pyinstrument>=5.1.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.14.14",
    "ty>=0.0.29",
    "types-python-dateutil>=2.9.0.20260402",
    "types-pyyaml>=6.0.12.20260408",
    "types-requests>=2.33.0.20260402",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchaudio = [{ index = "pytorch-cpu" }]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.logfire]
ignore_no_config = true

[tool.mypy]
check_untyped_defs = true

[tool.pytest]
addopts = ["-v", "--durations=0", "--cov=app", "--cov=agent_penny"]
env_files = [".env", ".test.env"]

[tool.ruff.lint]
extend-select = [
    # isort
    "I",
]

[tool.ty.src]
exclude = ["tests"]
